Rename linting/core package to linting/corelinting
