+++ internal/linting/corelinting/core_test.go
@@ -3,7 +3,7 @@
 // Copyright (c) 2023 HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package core
+package corelinting
 
 import (
 	"fmt"
