+++ internal/linting/corelinting/variable_with_no_type.go
@@ -3,7 +3,7 @@
 // Copyright (c) 2023 HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package core
+package corelinting
 
 import (
 	"context"
